pyproject.toml
[project]
name = "my-sdv-project"
version = "0.1.0"
description = "Battery Health Monitoring System using MQTT and Machine Learning"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "scipy>=1.17.0",
    "scikit-learn>=1.8.0",
    "joblib>=1.5.3",
    "paho-mqtt>=2.1.0",
]
